+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.0] - 2018-06-04
 ### Added
 - `IntoMutSliceLikePtr` trait for reinterpreting pointer types to compatible
   `SliceLike` pointers. This replaces the `IntoSliceLikeAllocation` trait for
@@ -157,7 +157,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/okready/scratchpad/compare/v1.0.0-beta.2...HEAD
+[1.0.0]: https://github.com/okready/scratchpad/compare/v1.0.0-beta.2...v1.0.0
 [1.0.0-beta.2]: https://github.com/okready/scratchpad/compare/v1.0.0-beta.1...v1.0.0-beta.2
 [1.0.0-beta.1]: https://github.com/okready/scratchpad/compare/0.3.0...v1.0.0-beta.1
 [0.3.0]: https://github.com/okready/scratchpad/compare/0.2.0...0.3.0